+++ tests/test_function/test_parse_string.py
@@ -201,3 +201,42 @@ def test_parse_unique_string() -> None:
 
     # then
     assert_that(result).is_equal_to(expected_result)
+
+
+def test_parse_uri() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "uri"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
+def test_parse_uri_component() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "uri_component"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)
+
+
+def test_parse_uri_component_to_string() -> None:
+    # given
+    sub_dir_path = EXAMPLES_DIR / "uri_component_to_string"
+    file_path = sub_dir_path / "main.bicep"
+    expected_result = json.loads((sub_dir_path / "result.json").read_text())
+
+    # when
+    result = BicepParser(file_path).json()
+
+    # then
+    assert_that(result).is_equal_to(expected_result)