+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support `last` function
 - Support `bool` function
 - Support `newGuid` function
+- Support `uri` function
+- Support `uriComponent` function
+- Support `uriComponentToString` function
 
 ## 0.1.0 - 2022-02-06